Login & JWT Token
login is now working, using username and password created when they sign up, i.e., user account is listed on the backend Db.
As well as this, request for tokens are being requested, received, and stored on the app.
continuing on to work on retrieving data from the backend db, and displaying them on the app.